# Optimized package.json + Add Jest and Jenkins CI