--- a/__tests__/addition.test.js
+++ b/__tests__/addition.test.js
@@ -1,0 +1,3 @@
+test('adds 1 + 2 to equal 3', () => {
+    expect(1 + 2).toBe(3)
+});