--- a/__tests__/division.test.js
+++ b/__tests__/division.test.js
@@ -1,0 +1,3 @@
+test('10 divided by 5 should be 2', () => {
+    expect(10 / 5).toBe(2);
+});