--- a/__tests__/multiplication.test.js
+++ b/__tests__/multiplication.test.js
@@ -1,0 +1,3 @@
+test('5 times 10 should be 50', () => {
+    expect(5 * 10).toBe(50);
+});