--- a/__tests__/substraction.test.js
+++ b/__tests__/substraction.test.js
@@ -1,0 +1,3 @@
+test('1 minus 3 should be -2', () => {
+    expect(1 - 3).toBe(-2);
+});