Optimized package.json to significantly speed up `yarn install`, the optimization yields on average (10 runs) a 36% speed increase (10.4s->6.6s) on debian, and 27% speed increase (29.4s->21.3s) on OSX. The debian machine is a retired server, while the OSX machine is a 2013 macbook air.

This is NOT a breaking change as long as you run `yarn install` as the readme states.